Rename runFixtureTests.js to run-fixture-tests.js.
Reviewed by @tolmasky.
