+++ packages/babel-parser/test/attachComment-false.js
@@ -1,5 +1,5 @@
 import path from "path";
-import runFixtureTests from "./helpers/runFixtureTests.js";
+import runFixtureTests from "./helpers/run-fixture-tests.js";
 import { parseExpression } from "../lib/index.js";
 import { fileURLToPath } from "url";
